#3 検索エリアの調整
